% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_lookup.R
\name{lookup_orcid}
\alias{lookup_orcid}
\title{Lookup orcid(s)}
\usage{
lookup_orcid(orcid)
}
\arguments{
\item{orcid}{character vector of orcid}
}
\value{
list of two tibbles with names and associated institutions
}
\description{
Lookup orcid(s)
}
