% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_lists.R
\name{nor_journals}
\alias{nor_journals}
\title{Norwegian journals from https://kanalregister.hkdir.no/}
\usage{
nor_journals()
}
\value{
tibble with results from web API call
}
\description{
Norwegian journals from https://kanalregister.hkdir.no/
}
