% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_lists.R
\name{nor_publishers}
\alias{nor_publishers}
\title{Norwegian publishers from https://kanalregister.hkdir.no/}
\usage{
nor_publishers()
}
\value{
tibble with results from web API call
}
\description{
Norwegian publishers from https://kanalregister.hkdir.no/
}
