% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_lookup.R
\name{orcid_lookup}
\alias{orcid_lookup}
\title{Lookup ORCiD}
\usage{
orcid_lookup(orcid)
}
\arguments{
\item{orcid}{the orcid identifier as a string}
}
\value{
a list of tibbles with information from the lookup
}
\description{
Lookup ORCiD
}
