% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_swecris_export.R
\encoding{UTF-8}
\name{parse_involved_people}
\alias{parse_involved_people}
\title{Parse "InvolvedPeople" field from a SweCRIS export}
\usage{
parse_involved_people(x)
}
\arguments{
\item{x}{the InvolvedPeople string}
}
\description{
InvolvedPeople is a field where strings are provided like
}
\details{
¤¤¤Peter Hedstroem¤0000-0003-1102-4342¤Principal Investigator¤Projektledare¤Male

These strings encode a dataset equivalent to what a call using \code{swecris_project_people("2021-00157_VR")}
would provide.

Instead of making multiple API request calls to resolve, a utility fcn for parsing is provided.
}
\examples{
\dontrun{
if(interactive()){
 ip <- swecris_kth$InvolvedPeople
 ppl <- purrr::map_dfr(ip, parse_involved_people, .id = "row")
 ppl$ProjectId <- swecris_kth[as.integer(ppl$row),]$ProjectId
 }
}
}
