% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_swecris_export.R
\encoding{UTF-8}
\name{parse_scb_codes}
\alias{parse_scb_codes}
\title{Parse "Scbs" field from a SweCRIS export}
\usage{
parse_scb_codes(x)
}
\arguments{
\item{x}{the Scbs string}
}
\description{
Scbs is a field where strings are provided like
}
\details{
¤¤¤ 1: Naturvetenskap, Natural Sciences, 103: Fysik, Physical Sciences, 10399: Annan fysik, Other Physics Topics ¤¤¤ 2: Teknik, Engineering and Technology, 205: Materialteknik, Materials Engineering, 20599: Annan materialteknik, Other Materials Engineering

These strings encode a dataset roughly equivalent to what a call using \code{swecris_project_scbs("2021-00157_VR")}
would provide

Instead of making multiple API request calls to resolve, a utility fcn for parsing is provided.
}
\examples{
\dontrun{
if(interactive()){
 x <- swecris_kth$Scbs
 scb <- purrr::map_dfr(x, parse_scb_codes, .id = "row")
 scb$ProjectId <- swecris_kth[as.integer(scb$row),]$ProjectId
 }
}
}
