% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_funders}
\alias{swecris_funders}
\title{Funders}
\usage{
swecris_funders()
}
\value{
a tibble
}
\description{
Funders in as a tibble
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
