% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_funding}
\alias{swecris_funding}
\title{Search the SweCRIS  API for funding data}
\usage{
swecris_funding(searchstring = "KTH, Kungliga Tekniska Högskolan", token)
}
\arguments{
\item{searchstring}{string to search for, Default is the search string for KTH}

\item{token}{token to be used for authentication}
}
\value{
an R object with the search result
}
\description{
Performs a search for funding and exports
results as an R object
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
