% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_fundings}
\alias{swecris_fundings}
\title{Fundings}
\usage{
swecris_fundings()
}
\value{
a tibble
}
\description{
Fundings in SweCRIS, more than 190k records of fundings for projects
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
