% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swecris_kth}
\alias{swecris_kth}
\title{SweCRIS data for KTH Royal Institute of Technology}
\format{
A list with two slots corresponding to
data returned from the betasearch API
\describe{
\item{hits}{projects - a nested object}
\item{tags}{subject category tags for projects}
}
}
\usage{
swecris_kth
}
\description{
A static dataset containing the projects in SweCRIS resulting from
a swecris_search() call on Fri 06 Nov 2020.
}
\keyword{datasets}
