% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_lists.R
\name{swecris_list_danish}
\alias{swecris_list_danish}
\title{Danish BFI lists}
\usage{
swecris_list_danish()
}
\value{
tibble with set column indicating type of list
(serier, forlag, fjerdene, tilfojede)
}
\description{
Danish BFI lists
}
\details{
see \href{https://ufm.dk/forskning-og-innovation/statistik-og-analyser/den-bibliometriske-forskningsindikator/BFI-lister}{details about available data}
}
