% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_lists.R
\name{swecris_list_finnish}
\alias{swecris_list_finnish}
\title{Finnish List}
\usage{
swecris_list_finnish()
}
\value{
a tibble
}
\description{
Finnish List of publication sources.
}
\details{
see \href{https://www.tsv.fi/julkaisufoorumi/haku.php?lang=en}{details about available data}
}
