% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swecris_list_norwegian}
\alias{swecris_list_norwegian}
\title{Norwegian list}
\format{
A table with data returned from the API
\describe{
\item{journal_id}{}
\item{title}{}
\item{title_en}{}
\item{issn_print}{}
\item{issn_online}{}
\item{oa}{}
\item{publishing_agreement}{}
\item{group_area}{}
\item{group_field}{}
\item{level_2026}{}
\item{level_2025}{}
\item{level_2024}{}
\item{level_2023}{}
\item{level_2022}{}
\item{level_2021}{}
\item{level_2020}{}
\item{level_2019}{}
\item{level_2018}{}
\item{level_2017}{}
\item{level_2016}{}
\item{level_2015}{}
\item{level_2014}{}
\item{level_2013}{}
\item{level_2012}{}
\item{level_2011}{}
\item{level_2010}{}
\item{level_2009}{}
\item{level_2008}{}
\item{level_2007}{}
\item{level_2006}{}
\item{level_2005}{}
\item{level_2004}{}
\item{publisher_id}{}
\item{publisher_company}{}
\item{publisher}{}
\item{publisher_country}{}
\item{language}{}
\item{conference_report}{}
\item{series}{}
\item{established}{}
\item{discontinued}{}
\item{url}{}
\item{last_updated}{}
\item{set}{}
\item{isbn_prefix}{}
\item{country}{}
}
}
\usage{
swecris_list_norwegian
}
\description{
A static dataset containing the Norwegian lists fetched 7 Aug 2023.
}
\details{
For details see these URLs:
\itemize{
\item https://dbh.nsd.uib.no/publiseringskanaler/AlltidFerskListeTidsskriftSomCsv
\item https://dbh.nsd.uib.no/publiseringskanaler/AlltidFerskListeForlagSomCsv
}
}
\keyword{datasets}
