% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic_lists.R
\name{swecris_list_swedish}
\alias{swecris_list_swedish}
\title{Swedish List}
\usage{
swecris_list_swedish()
}
\value{
a tibble
}
\description{
Swedish List of publication sources.
}
\details{
see \href{https://www.vr.se/uppdrag/oppen-vetenskap/svenska-listan---sakkunniggranskade-kanaler-i-swepub.html}{details about available data}
}
