% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_organisations}
\alias{swecris_organisations}
\title{Organizations}
\usage{
swecris_organisations()
}
\value{
a tibble
}
\description{
Organizations in SweCRIS
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
