% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_persons}
\alias{swecris_persons}
\title{Persons}
\usage{
swecris_persons(orgid)
}
\arguments{
\item{orgid}{optional orgid to filter for}
}
\value{
a tibble
}
\description{
Persons in SweCRIS
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
