% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_project}
\alias{swecris_project}
\title{Retrieve details about a project}
\usage{
swecris_project(project_id, format = c("tbl", "object"))
}
\arguments{
\item{project_id}{string identifier for the project}

\item{format}{one of "object" or "tbl"}
}
\value{
R object or tbl with results
}
\description{
Given the project identifier, retrieve details about a project.
}
