% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_project_leaders}
\alias{swecris_project_leaders}
\title{Retrieve project leader data}
\usage{
swecris_project_leaders(project_id)
}
\arguments{
\item{project_id}{string identifier for project}
}
\value{
R object with results
}
\description{
Given the project identifier, retrieve details about a project leaders
for a project.
}
