% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_project_people}
\alias{swecris_project_people}
\title{Retrieve project people data}
\usage{
swecris_project_people(project_id)
}
\arguments{
\item{project_id}{string identifier for project}
}
\value{
R data frame with results
}
\description{
Given the project identifier, retrieve details about a people involved with
a project.
}
\examples{
\dontrun{
if(interactive()){
 "2021-00157_VR" \%>\% swecris_project_people()
 }
}
}
