% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_project_scbs}
\alias{swecris_project_scbs}
\title{Retrieve project SCB codes data}
\usage{
swecris_project_scbs(project_id)
}
\arguments{
\item{project_id}{string identifier for project}
}
\value{
R data frame with results
}
\description{
Given the project identifier, retrieve details about subject matter codes.
}
\examples{
\dontrun{
if(interactive()){
 "2021-00157_VR" |> swecris_project_scbs()
 }
}
}
