% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_projects}
\alias{swecris_projects}
\title{Projects}
\usage{
swecris_projects(orgid)
}
\arguments{
\item{orgid}{orgid to filter for}
}
\value{
a tibble
}
\description{
Projects in SweCRIS
}
\details{
L

see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
