% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_projects_from_orcid}
\alias{swecris_projects_from_orcid}
\title{Projects data from ORCiD}
\usage{
swecris_projects_from_orcid(orcid)
}
\arguments{
\item{orcid}{A character string with a valid ORCiD}
}
\value{
a list with three slots (projects, peopleList, scbs)
}
\description{
Given an ORCiD, this function retrieves information about related projects,
about subject classification codes (a.k.a scbs) and about
involved people (peopleList)
}
\examples{
\dontrun{
if(interactive()){
 o <- "0000-0003-1102-4342" |> swecris_projects_from_orcid()
 o$projects
 o |> purrr::pluck("peopleList")
 o$scbs
 }
}
}
