% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_scb}
\alias{swecris_scb}
\title{SCB lookup table}
\usage{
swecris_scb()
}
\value{
a tibble
}
\description{
SCB 5-letter codes lookup table
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
