% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_swedish_list}
\alias{swecris_swedish_list}
\title{Swedish List}
\usage{
swecris_swedish_list()
}
\value{
a tibble
}
\description{
Swedish List of publications.
}
\details{
see \href{https://www.vr.se/uppdrag/oppen-vetenskap/svenska-listan---sakkunniggranskade-kanaler-i-swepub.html}{details about available data}
}
