% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swecris_kth}
\alias{swecris_kth}
\title{SweCRIS data for KTH Royal Institute of Technology}
\format{
A table with data returned from the API
\describe{
\item{ProjectId}{}
\item{ProjectTitleSv}{}
\item{ProjectTitleEn}{}
\item{ProjectAbstractSv}{}
\item{ProjectAbstractEn}{}
\item{ProjectStartDate}{}
\item{ProjectEndDate}{}
\item{CoordinatingOrganisationId}{}
\item{CoordinatingOrganisationNameSv}{}
\item{CoordinatingOrganisationNameEn}{}
\item{CoordinatingOrganisationTypeOfOrganisationSv}{}
\item{CoordinatingOrganisationTypeOfOrganisationEn}{}
\item{FundingOrganisationId}{}
\item{FundingOrganisationNameSv}{}
\item{FundingOrganisationNameEn}{}
\item{FundingOrganisationTypeOfOrganisationSv}{}
\item{FundingOrganisationTypeOfOrganisationEn}{}
\item{FundingsSek}{}
\item{FundingYear}{}
\item{FundingStartDate}{}
\item{FundingEndDate}{}
\item{TypeOfAwardId}{}
\item{TypeOfAwardDescrSv}{}
\item{TypeOfAwardDescrEn}{}
\item{InvolvedPeople}{}
\item{Scbs}{}
}
}
\usage{
swecris_kth
}
\description{
A static dataset containing the projects in SweCRIS resulting from
a swecris_funding() call on Mon 12 Oct 2022.
}
\keyword{datasets}
