% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaapi.R
\name{swecris_search}
\alias{swecris_search}
\title{Search the SweCRIS  API for an organization}
\usage{
swecris_search(orgs = "KTH, Kungliga tekniska högskolan")
}
\arguments{
\item{orgs}{string to search for, Default: "KTH, Kungliga Tekniska Högskolan"}
}
\value{
an R object with the search result
}
\description{
Performs a search for organizatinons and exports
results as an R object
}
\details{
see \href{https://swecris-api.vr.se/index.html}{details about available data}
}
