% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{swecris_search}
\alias{swecris_search}
\title{Search the SweCRIS betasearch API for an organization}
\usage{
swecris_search(orgs = "KTH, Kungliga tekniska högskolan")
}
\arguments{
\item{orgs}{string to search for, Default: "KTH, Kungliga Tekniska Högskolan"}
}
\value{
an R object with the search result
}
\description{
Performs a search against the betasearch API and exports
results as an R object
}
\details{
see \href{https://www.swecris.se/about/#export2}{details about available data}
}
